Crate heed_types

Types that can be used to serialize and deserialize types inside databases.

Structs

I8

Encodable version of i8.

I16

Encodable version of i16.

I32

Encodable version of i32.

I64

Encodable version of i64.

I128

Encodable version of i128.

Lazy

Owns bytes that can be decoded on demand.

LazyDecode

Lazily decodes the data bytes.

SerdeBincode

Describes a type that is Serialize/Deserialize and uses bincode to do so.

SerdeJson

Describes a type that is Serialize/Deserialize and uses serde_json to do so.

U8

Encodable version of u8.

U16

Encodable version of u16.

U32

Encodable version of u32.

U64

Encodable version of u64.

U128

Encodable version of u128.

Enums

Bytes

Describes a byte slice [u8] that is totally borrowed and doesn’t depend on any memory alignment.

DecodeIgnore

A convenient struct made to ignore the type when decoding it.

Str

Describes a str.

Unit

Describes the unit () type.